requirements.txt
Pillow==2.1.0
Django==2.2
djangorestframework==3.9.4
markdown==3.1.1
django-filter==2.1.0
django-rest-auth==0.9.5
django-allauth==0.39.1
psycopg2==2.8.3
djangorestframework-simplejwt==4.3.0
django-rest-swagger==2.2.0
drf-yasg==1.16.0